+++ R/tramo_outliers.R
@@ -26,7 +26,7 @@ NULL
 tramo_outliers <- function(y, order = c(0L, 1L, 1L), seasonal = c(0L, 1L, 1L), mean = FALSE,
                            X = NULL, X.td = NULL, ao = TRUE, ls = TRUE, tc = FALSE, so = FALSE, cv = 0, ml = FALSE, clean = FALSE) {
     if (!is.ts(y)) {
-        stop("y must be a time series")
+        stop("y must be a time series", call. = FALSE)
     }
     if (!is.null(X.td)) {
         td <- rjd3toolkit::td(s = y, groups = X.td)