+++ R/zzz.R
@@ -2,10 +2,10 @@
 NULL
 
 .onLoad <- function(libname, pkgname) {
-    if (!requireNamespace("rjd3toolkit", quietly = TRUE)) stop("Loading rjd3 libraries failed")
+    if (!requireNamespace("rjd3toolkit", quietly = TRUE)) stop("Loading rjd3 libraries failed", call. = FALSE)
 
     result <- rJava::.jpackage(pkgname, lib.loc = libname)
-    if (!result) stop("Loading java packages failed")
+    if (!result) stop("Loading java packages failed", call. = FALSE)
 
     proto.dir <- system.file("proto", package = pkgname)
     RProtoBuf::readProtoFiles2(protoPath = proto.dir)